Moved stringCapture code from serial.h to separate file
